+++ packages/swapper/src/swappers/cow/CowSwapper.ts
@@ -18,15 +18,21 @@ import {
   TradeResult,
   TradeTxs
 } from '../../api'
-import { CowApprovalNeeded } from './CowApprovalNeeded/CowApprovalNeeded'
-import { CowApproveInfinite } from './CowApproveInfinite/CowApproveInfinite'
+import { cowApprovalNeeded } from './cowApprovalNeeded/cowApprovalNeeded'
+import { cowApproveInfinite } from './cowApproveInfinite/cowApproveInfinite'
+import { getCowSwapTradeQuote } from './getCowSwapTradeQuote/getCowSwapTradeQuote'
 import { COWSWAP_UNSUPPORTED_ASSETS } from './utils/blacklist'
 import { getUsdRate } from './utils/helpers/helpers'
 
+/**
+ * CowSwap only supports ERC-20 swaps, hence ETH is not supported
+ * In order to get rates correctly, we need WETH asset to be passed as feeAsset
+ */
 export type CowSwapperDeps = {
   apiUrl: string
   adapter: ethereum.ChainAdapter
   web3: Web3
+  feeAsset: Asset // should be WETH asset
 }
 
 export class CowSwapper implements Swapper<'eip155:1'> {
@@ -36,9 +42,6 @@ export class CowSwapper implements Swapper<'eip155:1'> {
   constructor(deps: CowSwapperDeps) {
     this.deps = deps
   }
-
-  // eslint-disable-next-line @typescript-eslint/no-empty-function
-  async initialize() {}
 
   getType() {
     return SwapperType.CowSwap
@@ -50,8 +53,7 @@ export class CowSwapper implements Swapper<'eip155:1'> {
   }
 
   async getTradeQuote(input: GetTradeQuoteInput): Promise<TradeQuote<'eip155:1'>> {
-    console.info(input)
-    throw new Error('CowSwapper: getTradeQuote unimplemented')
+    return getCowSwapTradeQuote(this.deps, input)
   }
 
   async getUsdRate(input: Asset): Promise<string> {
@@ -64,11 +66,11 @@ export class CowSwapper implements Swapper<'eip155:1'> {
   }
 
   async approvalNeeded(args: ApprovalNeededInput<'eip155:1'>): Promise<ApprovalNeededOutput> {
-    return CowApprovalNeeded(this.deps, args)
+    return cowApprovalNeeded(this.deps, args)
   }
 
   async approveInfinite(args: ApproveInfiniteInput<'eip155:1'>): Promise<string> {
-    return CowApproveInfinite(this.deps, args)
+    return cowApproveInfinite(this.deps, args)
   }
 
   filterBuyAssetsBySellAssetId(args: BuyAssetBySellIdInput): AssetId[] {